// src/app/shared/model/zmst-agency.model.ts
export interface ZmstAgencyModel {
	agencyId: String;
	agencyName: String;
	agencyAbbr: String;
	agencyType: String;
	stateId: String;
	stateName: String;
	serviceTypeId: String;
	address: String;
	isActive: String;
	priority: Number;
	boardRequestLetter: string;
	
	docFileName: string;
    docContentType: string;

}